Edit questions in README for clarity and consistency
I had a hard time understanding a few of the questions, and spent some
time looking through the history to understand when they were added and
how they evolved. It seems like some of them were phased differently
between ISSUE_TEMPLATE and README, and I'm not familiar enough with all
the terms to tell if the differences are meaningful or not. I assume
there will be follow-ups to improve anything I make inaccurate.